flask/requirements.txt
Scrapy==2.6.0
google-search-results
flask
psycopg2-binary
marshmallow
flask-cors
python-dotenv